requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
# sqlite3 is built into Python
pydantic==2.5.0
python-multipart==0.0.6
PyMuPDF==1.23.8
pikepdf==8.7.1
pytesseract==0.3.10
Pillow==10.1.0
spacy==3.7.2
transformers==4.36.0
torch==2.1.0
tensorflow==2.15.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
email-validator==2.2.0
email-reply-parser==0.5.12
dateparser==1.1.8
python-dateutil==2.8.2
regex==2023.10.3
pytz>=2024.1